+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/bean/entity/HibernateTopic.java
@@ -29,8 +29,8 @@ public class HibernateTopic implements Bean {
     private String remark;
 
     // -----------------------------------------------------------一对多-----------------------------------------------------------
-    @OneToMany(cascade = CascadeType.MERGE, targetEntity = HibernateRelation.class, mappedBy = "topic")
-    private Set<HibernateRelation> relations = new HashSet<>();
+    @OneToMany(cascade = CascadeType.MERGE, targetEntity = HibernateSenderRelation.class, mappedBy = "topic")
+    private Set<HibernateSenderRelation> senderRelations = new HashSet<>();
 
     public HibernateTopic() {
     }
@@ -69,12 +69,12 @@ public class HibernateTopic implements Bean {
         this.remark = remark;
     }
 
-    public Set<HibernateRelation> getRelations() {
-        return relations;
+    public Set<HibernateSenderRelation> getSenderRelations() {
+        return senderRelations;
     }
 
-    public void setRelations(Set<HibernateRelation> relations) {
-        this.relations = relations;
+    public void setSenderRelations(Set<HibernateSenderRelation> senderRelations) {
+        this.senderRelations = senderRelations;
     }
 
     @Override