+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/dao/SenderRelationDaoImpl.java
@@ -1,10 +1,10 @@
 package com.dwarfeng.notify.impl.dao;
 
-import com.dwarfeng.notify.impl.bean.entity.HibernateRelation;
-import com.dwarfeng.notify.impl.bean.entity.key.HibernateRelationKey;
-import com.dwarfeng.notify.stack.bean.entity.Relation;
-import com.dwarfeng.notify.stack.bean.entity.key.RelationKey;
-import com.dwarfeng.notify.stack.dao.RelationDao;
+import com.dwarfeng.notify.impl.bean.entity.HibernateSenderRelation;
+import com.dwarfeng.notify.impl.bean.entity.key.HibernateSenderRelationKey;
+import com.dwarfeng.notify.stack.bean.entity.SenderRelation;
+import com.dwarfeng.notify.stack.bean.entity.key.SenderRelationKey;
+import com.dwarfeng.notify.stack.dao.SenderRelationDao;
 import com.dwarfeng.subgrade.impl.dao.HibernateBatchBaseDao;
 import com.dwarfeng.subgrade.impl.dao.HibernateEntireLookupDao;
 import com.dwarfeng.subgrade.impl.dao.HibernatePresetLookupDao;
@@ -18,17 +18,17 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Repository
-public class RelationDaoImpl implements RelationDao {
+public class SenderRelationDaoImpl implements SenderRelationDao {
 
-    private final HibernateBatchBaseDao<RelationKey, HibernateRelationKey, Relation, HibernateRelation>
+    private final HibernateBatchBaseDao<SenderRelationKey, HibernateSenderRelationKey, SenderRelation, HibernateSenderRelation>
             batchBaseDao;
-    private final HibernateEntireLookupDao<Relation, HibernateRelation> entireLookupDao;
-    private final HibernatePresetLookupDao<Relation, HibernateRelation> presetLookupDao;
+    private final HibernateEntireLookupDao<SenderRelation, HibernateSenderRelation> entireLookupDao;
+    private final HibernatePresetLookupDao<SenderRelation, HibernateSenderRelation> presetLookupDao;
 
-    public RelationDaoImpl(
-            HibernateBatchBaseDao<RelationKey, HibernateRelationKey, Relation, HibernateRelation> batchBaseDao,
-            HibernateEntireLookupDao<Relation, HibernateRelation> entireLookupDao,
-            HibernatePresetLookupDao<Relation, HibernateRelation> presetLookupDao
+    public SenderRelationDaoImpl(
+            HibernateBatchBaseDao<SenderRelationKey, HibernateSenderRelationKey, SenderRelation, HibernateSenderRelation> batchBaseDao,
+            HibernateEntireLookupDao<SenderRelation, HibernateSenderRelation> entireLookupDao,
+            HibernatePresetLookupDao<SenderRelation, HibernateSenderRelation> presetLookupDao
     ) {
         this.batchBaseDao = batchBaseDao;
         this.entireLookupDao = entireLookupDao;
@@ -38,35 +38,35 @@ public class RelationDaoImpl implements RelationDao {
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public RelationKey insert(Relation element) throws DaoException {
+    public SenderRelationKey insert(SenderRelation element) throws DaoException {
         return batchBaseDao.insert(element);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public void update(Relation element) throws DaoException {
+    public void update(SenderRelation element) throws DaoException {
         batchBaseDao.update(element);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public void delete(RelationKey key) throws DaoException {
+    public void delete(SenderRelationKey key) throws DaoException {
         batchBaseDao.delete(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public boolean exists(RelationKey key) throws DaoException {
+    public boolean exists(SenderRelationKey key) throws DaoException {
         return batchBaseDao.exists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public Relation get(RelationKey key) throws DaoException {
+    public SenderRelation get(SenderRelationKey key) throws DaoException {
         return batchBaseDao.get(key);
     }
 
@@ -74,35 +74,35 @@ public class RelationDaoImpl implements RelationDao {
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public List<RelationKey> batchInsert(@SkipRecord List<Relation> elements) throws DaoException {
+    public List<SenderRelationKey> batchInsert(@SkipRecord List<SenderRelation> elements) throws DaoException {
         return batchBaseDao.batchInsert(elements);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public void batchUpdate(@SkipRecord List<Relation> elements) throws DaoException {
+    public void batchUpdate(@SkipRecord List<SenderRelation> elements) throws DaoException {
         batchBaseDao.batchUpdate(elements);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public void batchDelete(@SkipRecord List<RelationKey> keys) throws DaoException {
+    public void batchDelete(@SkipRecord List<SenderRelationKey> keys) throws DaoException {
         batchBaseDao.batchDelete(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public boolean allExists(@SkipRecord List<RelationKey> keys) throws DaoException {
+    public boolean allExists(@SkipRecord List<SenderRelationKey> keys) throws DaoException {
         return batchBaseDao.allExists(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public boolean nonExists(@SkipRecord List<RelationKey> keys) throws DaoException {
+    public boolean nonExists(@SkipRecord List<SenderRelationKey> keys) throws DaoException {
         return batchBaseDao.nonExists(keys);
     }
 
@@ -110,7 +110,7 @@ public class RelationDaoImpl implements RelationDao {
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public List<Relation> batchGet(@SkipRecord List<RelationKey> keys) throws DaoException {
+    public List<SenderRelation> batchGet(@SkipRecord List<SenderRelationKey> keys) throws DaoException {
         return batchBaseDao.batchGet(keys);
     }
 
@@ -118,7 +118,7 @@ public class RelationDaoImpl implements RelationDao {
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public List<Relation> lookup() throws DaoException {
+    public List<SenderRelation> lookup() throws DaoException {
         return entireLookupDao.lookup();
     }
 
@@ -126,7 +126,7 @@ public class RelationDaoImpl implements RelationDao {
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public List<Relation> lookup(PagingInfo pagingInfo) throws DaoException {
+    public List<SenderRelation> lookup(PagingInfo pagingInfo) throws DaoException {
         return entireLookupDao.lookup(pagingInfo);
     }
 
@@ -141,7 +141,7 @@ public class RelationDaoImpl implements RelationDao {
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public List<Relation> lookup(String preset, Object[] objs) throws DaoException {
+    public List<SenderRelation> lookup(String preset, Object[] objs) throws DaoException {
         return presetLookupDao.lookup(preset, objs);
     }
 
@@ -149,7 +149,7 @@ public class RelationDaoImpl implements RelationDao {
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public List<Relation> lookup(String preset, Object[] objs, PagingInfo pagingInfo) throws DaoException {
+    public List<SenderRelation> lookup(String preset, Object[] objs, PagingInfo pagingInfo) throws DaoException {
         return presetLookupDao.lookup(preset, objs, pagingInfo);
     }