+++ notify-impl/src/test/java/com/dwarfeng/notify/impl/configuration/ServiceConfiguration.java
@@ -4,7 +4,7 @@ import com.dwarfeng.notify.impl.service.operation.NotifySettingCrudOperation;
 import com.dwarfeng.notify.impl.service.operation.SenderInfoCrudOperation;
 import com.dwarfeng.notify.impl.service.operation.TopicCrudOperation;
 import com.dwarfeng.notify.stack.bean.entity.*;
-import com.dwarfeng.notify.stack.bean.entity.key.RelationKey;
+import com.dwarfeng.notify.stack.bean.entity.key.SenderRelationKey;
 import com.dwarfeng.notify.stack.cache.*;
 import com.dwarfeng.notify.stack.dao.*;
 import com.dwarfeng.sfds.api.integration.subgrade.SnowFlakeLongIdKeyFetcher;
@@ -40,8 +40,8 @@ public class ServiceConfiguration {
     private final SenderSupportCache senderSupportCache;
     private final TopicCrudOperation topicCrudOperation;
     private final TopicDao topicDao;
-    private final RelationDao relationDao;
-    private final RelationCache relationCache;
+    private final SenderRelationDao senderRelationDao;
+    private final SenderRelationCache senderRelationCache;
 
     @Value("${cache.timeout.entity.user}")
     private long userTimeout;
@@ -51,8 +51,8 @@ public class ServiceConfiguration {
     private long routerSupportTimeout;
     @Value("${cache.timeout.entity.sender_support}")
     private long senderSupportTimeout;
-    @Value("${cache.timeout.entity.relation}")
-    private long relationTimeout;
+    @Value("${cache.timeout.entity.sender_relation}")
+    private long senderRelationTimeout;
 
     public ServiceConfiguration(
             ServiceExceptionMapperConfiguration serviceExceptionMapperConfiguration,
@@ -63,7 +63,7 @@ public class ServiceConfiguration {
             SenderInfoCrudOperation senderInfoCrudOperation, SenderInfoDao senderInfoDao,
             SenderSupportDao senderSupportDao, SenderSupportCache senderSupportCache,
             TopicCrudOperation topicCrudOperation, TopicDao topicDao,
-            RelationDao relationDao, RelationCache relationCache
+            SenderRelationDao senderRelationDao, SenderRelationCache senderRelationCache
     ) {
         this.serviceExceptionMapperConfiguration = serviceExceptionMapperConfiguration;
         this.userDao = userDao;
@@ -80,8 +80,8 @@ public class ServiceConfiguration {
         this.senderSupportCache = senderSupportCache;
         this.topicCrudOperation = topicCrudOperation;
         this.topicDao = topicDao;
-        this.relationDao = relationDao;
-        this.relationCache = relationCache;
+        this.senderRelationDao = senderRelationDao;
+        this.senderRelationCache = senderRelationCache;
     }
 
     @Bean
@@ -285,30 +285,30 @@ public class ServiceConfiguration {
     }
 
     @Bean
-    public GeneralBatchCrudService<RelationKey, Relation> relationGeneralBatchCrudService() {
+    public GeneralBatchCrudService<SenderRelationKey, SenderRelation> senderRelationGeneralBatchCrudService() {
         return new GeneralBatchCrudService<>(
-                relationDao,
-                relationCache,
+                senderRelationDao,
+                senderRelationCache,
                 new ExceptionKeyFetcher<>(),
                 serviceExceptionMapperConfiguration.mapServiceExceptionMapper(),
                 LogLevel.WARN,
-                relationTimeout
+                senderRelationTimeout
         );
     }
 
     @Bean
-    public DaoOnlyEntireLookupService<Relation> relationDaoOnlyEntireLookupService() {
+    public DaoOnlyEntireLookupService<SenderRelation> senderRelationDaoOnlyEntireLookupService() {
         return new DaoOnlyEntireLookupService<>(
-                relationDao,
+                senderRelationDao,
                 serviceExceptionMapperConfiguration.mapServiceExceptionMapper(),
                 LogLevel.WARN
         );
     }
 
     @Bean
-    public DaoOnlyPresetLookupService<Relation> relationDaoOnlyPresetLookupService() {
+    public DaoOnlyPresetLookupService<SenderRelation> senderRelationDaoOnlyPresetLookupService() {
         return new DaoOnlyPresetLookupService<>(
-                relationDao,
+                senderRelationDao,
                 serviceExceptionMapperConfiguration.mapServiceExceptionMapper(),
                 LogLevel.WARN
         );