+++ notify-sdk/src/main/java/com/dwarfeng/notify/sdk/bean/entity/key/FastJsonSenderRelationKey.java
@@ -1,26 +1,26 @@
 package com.dwarfeng.notify.sdk.bean.entity.key;
 
 import com.alibaba.fastjson.annotation.JSONField;
-import com.dwarfeng.notify.stack.bean.entity.key.RelationKey;
+import com.dwarfeng.notify.stack.bean.entity.key.SenderRelationKey;
 import com.dwarfeng.subgrade.stack.bean.key.Key;
 
 import java.util.Objects;
 
 /**
- * FastJson 关系主键。
+ * FastJson 发送器关系主键。
  *
  * @author DwArFeng
  * @since 1.0.0
  */
-public class FastJsonRelationKey implements Key {
+public class FastJsonSenderRelationKey implements Key {
 
-    private static final long serialVersionUID = 4071286139261304477L;
+    private static final long serialVersionUID = -3978239476963933375L;
 
-    public static FastJsonRelationKey of(RelationKey relationKey) {
-        if (Objects.isNull(relationKey)) {
+    public static FastJsonSenderRelationKey of(SenderRelationKey senderRelationKey) {
+        if (Objects.isNull(senderRelationKey)) {
             return null;
         } else {
-            return new FastJsonRelationKey(relationKey.getNotifySettingId(), relationKey.getTopicId());
+            return new FastJsonSenderRelationKey(senderRelationKey.getNotifySettingId(), senderRelationKey.getTopicId());
         }
     }
 
@@ -30,10 +30,10 @@ public class FastJsonRelationKey implements Key {
     @JSONField(name = "topic_id", ordinal = 2)
     private String topicId;
 
-    public FastJsonRelationKey() {
+    public FastJsonSenderRelationKey() {
     }
 
-    public FastJsonRelationKey(Long notifySettingId, String topicId) {
+    public FastJsonSenderRelationKey(Long notifySettingId, String topicId) {
         this.notifySettingId = notifySettingId;
         this.topicId = topicId;
     }
@@ -59,7 +59,7 @@ public class FastJsonRelationKey implements Key {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
 
-        FastJsonRelationKey that = (FastJsonRelationKey) o;
+        FastJsonSenderRelationKey that = (FastJsonSenderRelationKey) o;
 
         if (!Objects.equals(notifySettingId, that.notifySettingId)) return false;
         return Objects.equals(topicId, that.topicId);
@@ -74,7 +74,7 @@ public class FastJsonRelationKey implements Key {
 
     @Override
     public String toString() {
-        return "FastJsonRelationKey{" +
+        return "FastJsonSenderRelationKey{" +
                 "notifySettingId=" + notifySettingId +
                 ", topicId='" + topicId + '\'' +
                 '}';