+++ notify-sdk/src/main/java/com/dwarfeng/notify/sdk/bean/entity/key/JSFixedFastJsonSenderRelationKey.java
@@ -1,26 +1,40 @@
-package com.dwarfeng.notify.stack.bean.entity.key;
+package com.dwarfeng.notify.sdk.bean.entity.key;
 
+import com.alibaba.fastjson.annotation.JSONField;
+import com.alibaba.fastjson.serializer.ToStringSerializer;
+import com.dwarfeng.notify.stack.bean.entity.key.SenderRelationKey;
 import com.dwarfeng.subgrade.stack.bean.key.Key;
 
 import java.util.Objects;
 
 /**
- * 关系主键。
+ * JSFixed FastJson 发送器关系主键。
  *
  * @author DwArFeng
  * @since 1.0.0
  */
-public class RelationKey implements Key {
+public class JSFixedFastJsonSenderRelationKey implements Key {
 
-    private static final long serialVersionUID = -2815133004629979563L;
+    private static final long serialVersionUID = 1386025715592420422L;
 
-    private Long notifySettingId;
-    private String topicId;
-
-    public RelationKey() {
+    public static JSFixedFastJsonSenderRelationKey of(SenderRelationKey senderRelationKey) {
+        if (Objects.isNull(senderRelationKey)) {
+            return null;
+        } else {
+            return new JSFixedFastJsonSenderRelationKey(senderRelationKey.getNotifySettingId(), senderRelationKey.getTopicId());
+        }
     }
 
-    public RelationKey(Long notifySettingId, String topicId) {
+    @JSONField(name = "notify_setting_id", ordinal = 1, serializeUsing = ToStringSerializer.class)
+    private Long notifySettingId;
+
+    @JSONField(name = "topic_id", ordinal = 2)
+    private String topicId;
+
+    public JSFixedFastJsonSenderRelationKey() {
+    }
+
+    public JSFixedFastJsonSenderRelationKey(Long notifySettingId, String topicId) {
         this.notifySettingId = notifySettingId;
         this.topicId = topicId;
     }
@@ -46,10 +60,9 @@ public class RelationKey implements Key {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
 
-        RelationKey that = (RelationKey) o;
+        JSFixedFastJsonSenderRelationKey that = (JSFixedFastJsonSenderRelationKey) o;
 
-        if (!Objects.equals(notifySettingId, that.notifySettingId))
-            return false;
+        if (!Objects.equals(notifySettingId, that.notifySettingId)) return false;
         return Objects.equals(topicId, that.topicId);
     }
 
@@ -62,7 +75,7 @@ public class RelationKey implements Key {
 
     @Override
     public String toString() {
-        return "RelationKey{" +
+        return "JSFixedFastJsonSenderRelationKey{" +
                 "notifySettingId=" + notifySettingId +
                 ", topicId='" + topicId + '\'' +
                 '}';