+++ notify-stack/src/main/java/com/dwarfeng/notify/stack/bean/entity/key/SenderRelationKey.java
@@ -1,40 +1,26 @@
-package com.dwarfeng.notify.sdk.bean.entity.key;
+package com.dwarfeng.notify.stack.bean.entity.key;
 
-import com.alibaba.fastjson.annotation.JSONField;
-import com.alibaba.fastjson.serializer.ToStringSerializer;
-import com.dwarfeng.notify.stack.bean.entity.key.RelationKey;
 import com.dwarfeng.subgrade.stack.bean.key.Key;
 
 import java.util.Objects;
 
 /**
- * JSFixed FastJson 关系主键。
+ * 发送器关系主键。
  *
  * @author DwArFeng
  * @since 1.0.0
  */
-public class JSFixedFastJsonRelationKey implements Key {
+public class SenderRelationKey implements Key {
 
-    private static final long serialVersionUID = 6484470269472125364L;
+    private static final long serialVersionUID = -8949434525933907828L;
 
-    public static JSFixedFastJsonRelationKey of(RelationKey relationKey) {
-        if (Objects.isNull(relationKey)) {
-            return null;
-        } else {
-            return new JSFixedFastJsonRelationKey(relationKey.getNotifySettingId(), relationKey.getTopicId());
-        }
-    }
-
-    @JSONField(name = "notify_setting_id", ordinal = 1, serializeUsing = ToStringSerializer.class)
     private Long notifySettingId;
-
-    @JSONField(name = "topic_id", ordinal = 2)
     private String topicId;
 
-    public JSFixedFastJsonRelationKey() {
+    public SenderRelationKey() {
     }
 
-    public JSFixedFastJsonRelationKey(Long notifySettingId, String topicId) {
+    public SenderRelationKey(Long notifySettingId, String topicId) {
         this.notifySettingId = notifySettingId;
         this.topicId = topicId;
     }
@@ -60,9 +46,10 @@ public class JSFixedFastJsonRelationKey implements Key {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
 
-        JSFixedFastJsonRelationKey that = (JSFixedFastJsonRelationKey) o;
+        SenderRelationKey that = (SenderRelationKey) o;
 
-        if (!Objects.equals(notifySettingId, that.notifySettingId)) return false;
+        if (!Objects.equals(notifySettingId, that.notifySettingId))
+            return false;
         return Objects.equals(topicId, that.topicId);
     }
 
@@ -75,7 +62,7 @@ public class JSFixedFastJsonRelationKey implements Key {
 
     @Override
     public String toString() {
-        return "JSFixedFastJsonRelationKey{" +
+        return "SenderRelationKey{" +
                 "notifySettingId=" + notifySettingId +
                 ", topicId='" + topicId + '\'' +
                 '}';