+++ CHANGELOG.md
@@ -14,7 +14,7 @@
   - com.dwarfeng.notify.stack.bean.entity.SenderInfo。
   - com.dwarfeng.notify.stack.bean.entity.SenderSupport。
   - com.dwarfeng.notify.stack.bean.entity.Topic。
-  - com.dwarfeng.notify.stack.bean.entity.Relation。
+  - com.dwarfeng.notify.stack.bean.entity.SenderRelation。
 
 #### Bug修复